Overhauled the Database and combined the transactions into one table. And changed all the corresponding file to the new database schema

+++ app/src/main/java/calpoly/edu/stracker/Transaction.java
@@ -9,7 +9,7 @@ public class Transaction {
     boolean isIncome;
     String date;
     String title;
-    int amount;
+    double amount;
     String category;
 
     public Transaction() {
@@ -48,11 +48,11 @@ public class Transaction {
         this.title = title;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
